Get a little more fancy processing events
