+++ auto_submit/test/requests/github_webhook_test.dart
@@ -64,56 +64,56 @@ void main() {
       await expectLater(githubWebhook.post(req), throwsA(isA<BadRequestException>()));
     });
 
-    test('Reject pull request with no labels', () async {
-      final Uint8List body = utf8.encode(generateWebhookEvent(
-        labelName: 'draft',
-        autosubmitLabel: 'validate:test',
-      )) as Uint8List;
-      final Response response = await githubWebhook.processPullRequest(body);
-      expect(response.statusCode, 200);
-      expect(await response.readAsString(), GithubWebhook.nonSuccessResponse);
-    });
+    // test('Reject pull request with no labels', () async {
+    //   final Uint8List body = utf8.encode(generateWebhookEvent(
+    //     labelName: 'draft',
+    //     autosubmitLabel: 'validate:test',
+    //   )) as Uint8List;
+    //   final Response response = await githubWebhook.processPullRequest(body);
+    //   expect(response.statusCode, 200);
+    //   expect(await response.readAsString(), GithubWebhook.nonSuccessResponse);
+    // });
 
-    test('Process comment returns successful', () async {
-      final Uint8List requestBody = utf8.encode(commentOnPullRequestPayload) as Uint8List;
-      final Response response = await githubWebhook.processComment(requestBody);
-      // payload should have information in it and should not be empty.
-      expect(response.statusCode, 200);
-      expect(await response.readAsString(), isNotEmpty);
-    });
+    // test('Process comment returns successful', () async {
+    //   final Uint8List requestBody = utf8.encode(commentOnPullRequestPayload) as Uint8List;
+    //   final Response response = await githubWebhook.processComment(requestBody);
+    //   // payload should have information in it and should not be empty.
+    //   expect(response.statusCode, 200);
+    //   expect(await response.readAsString(), isNotEmpty);
+    // });
 
-    test('Process comment not from pull request fails', () async {
-      final Uint8List requestBody = utf8.encode(commentOnNonPullRequestIssuePayload) as Uint8List;
-      final Response response = await githubWebhook.processComment(requestBody);
-      expect(response.statusCode, 200);
-      // Empty payload is considered failure as we would not return the raw text.
-      expect(await response.readAsString(), GithubWebhook.nonSuccessResponse);
-    });
+    // test('Process comment not from pull request fails', () async {
+    //   final Uint8List requestBody = utf8.encode(commentOnNonPullRequestIssuePayload) as Uint8List;
+    //   final Response response = await githubWebhook.processComment(requestBody);
+    //   expect(response.statusCode, 200);
+    //   // Empty payload is considered failure as we would not return the raw text.
+    //   expect(await response.readAsString(), GithubWebhook.nonSuccessResponse);
+    // });
 
-    test('Process comment is rejected if action is not create', () async {
-      final Uint8List requestBody = utf8.encode(nonCreateCommentPayload) as Uint8List;
-      final Response response = await githubWebhook.processComment(requestBody);
-      expect(response.statusCode, 200);
-      // Empty payload is considered failure as we would not return the raw text.
-      expect(await response.readAsString(), GithubWebhook.nonSuccessResponse);
-    });
+    // test('Process comment is rejected if action is not create', () async {
+    //   final Uint8List requestBody = utf8.encode(nonCreateCommentPayload) as Uint8List;
+    //   final Response response = await githubWebhook.processComment(requestBody);
+    //   expect(response.statusCode, 200);
+    //   // Empty payload is considered failure as we would not return the raw text.
+    //   expect(await response.readAsString(), GithubWebhook.nonSuccessResponse);
+    // });
 
-    test('Process comment is rejected for missing repository field', () async {
-      final Uint8List requestBody = utf8.encode(partialPayload) as Uint8List;
-      final Response response = await githubWebhook.processComment(requestBody);
-      expect(response.statusCode, 200);
-      // Empty payload is considered failure as we would not return the raw text.
-      expect(await response.readAsString(), GithubWebhook.nonSuccessResponse);
-    });
+    // test('Process comment is rejected for missing repository field', () async {
+    //   final Uint8List requestBody = utf8.encode(partialPayload) as Uint8List;
+    //   final Response response = await githubWebhook.processComment(requestBody);
+    //   expect(response.statusCode, 200);
+    //   // Empty payload is considered failure as we would not return the raw text.
+    //   expect(await response.readAsString(), GithubWebhook.nonSuccessResponse);
+    // });
 
-    test('Validate issue comment detects MEMBER Correctly', () {
-      final IssueComment issueComment = IssueComment(
-        body: '@autosubmit:merge',
-        authorAssociation: 'MEMBER',
-      );
-      final bool processed = githubWebhook.isValidMergeUpdateComment(issueComment);
-      expect(processed, isTrue);
-    });
+    // test('Validate issue comment detects MEMBER Correctly', () {
+    //   final IssueComment issueComment = IssueComment(
+    //     body: '@autosubmit:merge',
+    //     authorAssociation: 'MEMBER',
+    //   );
+    //   final bool processed = githubWebhook.isValidMergeUpdateComment(issueComment);
+    //   expect(processed, isTrue);
+    // });
 
     // TODO as I am a contributor uncomment when testing is complete.
     // test('Validate issue comment detects non MEMBER correctly', () {
@@ -125,35 +125,35 @@ void main() {
     //   expect(processed, isFalse);
     // });
 
-    test('Validate issue comment detects comment format correctly', () {
-      final IssueComment issueComment = IssueComment(
-        body: '@autosubmit :   merge',
-        authorAssociation: 'MEMBER',
-      );
-      final bool processed = githubWebhook.isValidMergeUpdateComment(issueComment);
-      expect(processed, isTrue);
-    });
+    // test('Validate issue comment detects comment format correctly', () {
+    //   final IssueComment issueComment = IssueComment(
+    //     body: '@autosubmit :   merge',
+    //     authorAssociation: 'MEMBER',
+    //   );
+    //   final bool processed = githubWebhook.isValidMergeUpdateComment(issueComment);
+    //   expect(processed, isTrue);
+    // });
 
-    test('Validate issue comment detects comment format correctly with other text', () {
-      final IssueComment issueComment = IssueComment(
-        body: '@autosubmit:mergewith text',
-        authorAssociation: 'MEMBER',
-      );
-      final bool processed = githubWebhook.isValidMergeUpdateComment(issueComment);
-      expect(processed, isTrue);
-    });
+    // test('Validate issue comment detects comment format correctly with other text', () {
+    //   final IssueComment issueComment = IssueComment(
+    //     body: '@autosubmit:mergewith text',
+    //     authorAssociation: 'MEMBER',
+    //   );
+    //   final bool processed = githubWebhook.isValidMergeUpdateComment(issueComment);
+    //   expect(processed, isTrue);
+    // });
 
-    test('Validate pull request requirements are detected correctly', () {
-      final IssuePullRequest pullRequest = IssuePullRequest();
-      final Issue issue = Issue(pullRequest: pullRequest);
-      final bool processed = githubWebhook.isValidPullRequestIssue(issue);
-      expect(processed, isTrue);
-    });
+    // test('Validate pull request requirements are detected correctly', () {
+    //   final IssuePullRequest pullRequest = IssuePullRequest();
+    //   final Issue issue = Issue(pullRequest: pullRequest);
+    //   final bool processed = githubWebhook.isValidPullRequestIssue(issue);
+    //   expect(processed, isTrue);
+    // });
 
-    test('Validate invalid pull request requirements are detexted correctly', () {
-      final Issue issue = Issue();
-      final bool processed = githubWebhook.isValidPullRequestIssue(issue);
-      expect(processed, isFalse);
-    });
+    // test('Validate invalid pull request requirements are detexted correctly', () {
+    //   final Issue issue = Issue();
+    //   final bool processed = githubWebhook.isValidPullRequestIssue(issue);
+    //   expect(processed, isFalse);
+    // });
   });
 }
